requirements-whisper.txt
# needs to be in a separate requirements bc poetry cannot lock a dependency for openai-whisper
openai-whisper
